Init create woo item

+++ inc/class-event-ajax.php
@@ -121,7 +121,7 @@ class Event_Ajax {
 			$payment = isset( $_POST['payment_method'] ) ? sanitize_text_field( $_POST['payment_method'] ) : false;
 
 			// create new book return $booking_id if success and WP Error if fail
-			$args = apply_filters( 'event_auth_create_booking_args', array(
+			$args = apply_filters( 'tp_event_create_booking_args', array(
 				'event_id'   => $event_id,
 				'qty'        => $qty,
 				'price'      => (float) $event->get_price() * $qty,
@@ -136,37 +136,42 @@ class Event_Ajax {
 				throw new Exception( sprintf( '%s %s', get_title(), __( 'is not ready. Please contact administrator to setup payment gateways.', 'tp-event' ) ) );
 			}
 
-			$booking_id = $booking->create_booking( $args );
-			// create booking result
-			if ( is_wp_error( $booking_id ) ) {
-				throw new Exception( $booking_id->get_error_message() );
-			} else {
-				if ( $args['price'] == 0 ) {
-					// update booking status
-					$book = Event_Booking::instance( $booking_id );
-					$book->update_status( 'pending' );
-
-					// user booking
-					$user = get_userdata( $book->user_id );
-					tp_event_add_notice( 'success', sprintf( __( 'Book ID <strong>%s</strong> completed! We\'ll send mail to <strong>%s</strong> when it is approve.', 'tp-event' ), tp_event_format_ID( $booking_id ), $user->user_email ) );
-					wp_send_json( apply_filters( 'event_auth_register_ajax_result', array(
-						'status' => true,
-						'url'    => tp_event_account_url()
-					) ) );
-				} else if ( $payment ) {
-					$return = $payment->process( $booking_id );
-
-					if ( isset( $return['status'] ) && $return['status'] === false ) {
-						wp_delete_post( $booking_id );
-					}
-
-					wp_send_json( $return );
+			if ( $args['payment_id'] == 'paypal' ) {
+				$booking_id = $booking->create_booking( $args );
+				// create booking result
+				if ( is_wp_error( $booking_id ) ) {
+					throw new Exception( $booking_id->get_error_message() );
 				} else {
-					wp_send_json( array(
-						'status'  => false,
-						'message' => __( 'Payment method is not available', 'tp-event' )
-					) );
+					if ( $args['price'] == 0 ) {
+						// update booking status
+						$book = Event_Booking::instance( $booking_id );
+						$book->update_status( 'pending' );
+
+						// user booking
+						$user = get_userdata( $book->user_id );
+						tp_event_add_notice( 'success', sprintf( __( 'Book ID <strong>%s</strong> completed! We\'ll send mail to <strong>%s</strong> when it is approve.', 'tp-event' ), tp_event_format_ID( $booking_id ), $user->user_email ) );
+						wp_send_json( apply_filters( 'event_auth_register_ajax_result', array(
+							'status' => true,
+							'url'    => tp_event_account_url()
+						) ) );
+					} else if ( $payment ) {
+						$return = $payment->process( $booking_id );
+
+						if ( isset( $return['status'] ) && $return['status'] === false ) {
+							wp_delete_post( $booking_id );
+						}
+
+						wp_send_json( $return );
+					} else {
+						wp_send_json( array(
+							'status'  => false,
+							'message' => __( 'Payment method is not available', 'tp-event' )
+						) );
+					}
 				}
+			} elseif ( $args['payment_id'] == 'woocommerce' ) {
+				$booking->add_to_woo_item($args);
+
 			}
 		} catch ( Exception $e ) {
 			if ( $e ) {
